refactor: Use template files instead of hardcoded dockerfile
