+++ internal/templates/rust-dockerfile.tmpl
@@ -1,0 +1,17 @@
+FROM rust:{{.Version}}-buster as builder
+
+WORKDIR /app
+
+COPY . .
+
+RUN cargo build --release
+
+FROM debian:buster-slim
+
+WORKDIR /usr/local/bin
+
+COPY --from=builder /app/target/release/{{.ProjectName}} .
+
+RUN apt-get update && apt install -y openssl
+
+CMD ["./{{.ProjectName}}"]
